+++ Changelog.md
@@ -40,6 +40,7 @@ types should preserve whitespace, while all other primitives have collapse behav
 - [#868]: Allow to have both `$text` and `$value` special fields in one struct. Previously
   any text will be recognized as `$value` field even when `$text` field is also presented.
 - [#868]: Skip text events when deserialize a sequence of items overlapped with text (including CDATA).
+- [#841]: Do not strip `xml` prefix from the attributes when map them to struct fields in `Deserializer`.
 
 ### Misc Changes
 
@@ -50,6 +51,7 @@ types should preserve whitespace, while all other primitives have collapse behav
 
 [#285]: https://github.com/tafia/quick-xml/issues/285
 [#766]: https://github.com/tafia/quick-xml/pull/766
+[#841]: https://github.com/tafia/quick-xml/issues/841
 [#863]: https://github.com/tafia/quick-xml/pull/863
 [#868]: https://github.com/tafia/quick-xml/pull/868
 [general entity]: https://www.w3.org/TR/xml11/#gen-entity